refactored App component
